native-messaging-app/requirements.txt
pyOpenSSL==24.1.0
cryptography==42.0.8